moved investment specific task logic to tasks module
