+++ src/client/tasks.js
@@ -361,8 +361,6 @@ import {
   TASK_GET_ONE_LIST_DETAILS,
   TASK_SAVE_ONE_LIST_DETAILS,
 } from './modules/Companies/CoreTeam/state'
-import { TASK_SAVE_INVESTMENT_PROJECT_TASK } from './modules/Investments/Projects/Tasks/state'
-import { saveInvestmentProjectTask } from './modules/Investments/Projects/Tasks/tasks'
 
 import {
   TASK_EDIT_OMIS_QUOTE_INFORMATION,
@@ -613,7 +611,6 @@ export const tasks = {
   [TASK_EDIT_OMIS_QUOTE_INFORMATION]: updateOrder,
   [TASK_EDIT_OMIS_INTERNAL_INFORMATION]: updateOrder,
   [TASK_GET_TASK_DETAILS]: getTaskDetail,
-  [TASK_SAVE_INVESTMENT_PROJECT_TASK]: saveInvestmentProjectTask,
   [TASK_GET_INVESTMENT_PROJECT]: investmentProjectTasks.getInvestmentProject,
   [TASK_ARCHIVE_TASK]: archiveTask,
   [TASK_RECONCILE_OMIS_PAYMENT]: savePayment,
