+++ test/functional/cypress/fakers/task.js
@@ -28,7 +28,6 @@ const taskFaker = (overrides = {}) => ({
   modifiedBy: basicAdviserFaker(),
   modifiedOn: faker.date.past().toISOString(),
   createdOn: faker.date.past().toISOString(),
-  investmentProjectTask: null,
 
   ...overrides,
 })
